Fixed the fetch in the quizbuilder.
Added JsonOutputFunctionsParser from Langchain to turn the string response from the AI model into a usable JSON object.

Added a loop to check if the AI's response is a valid JSON object. If it isn't, it will try up to 5 attempts of getting a valid response.

Added a temporary button for generating a question in the QuizBuilder.tsx

Removed AI_Test.js