36 als docent wil ik een vraag suggereert doormiddel van ai (#66)
* Setup Langchain and OpenAI.

Tested prompting the AI and extracting the JSON response.

* changes for nelio

* Base for implementing the AI.

* Added a route for the AI request

* Fixed the fetch in the quizbuilder.

Added JsonOutputFunctionsParser from Langchain to turn the string response from the AI model into a usable JSON object.

Added a loop to check if the AI's response is a valid JSON object. If it isn't, it will try up to 5 attempts of getting a valid response.

Added a temporary button for generating a question in the QuizBuilder.tsx

Removed AI_Test.js

* Added files containing the content of each information page.

Added a topic to content mapping.

Adjusted the AI system so that it gets the content for the corresponding topic. This helps it generate questions that fit our content page better.

Added a dropdown select for selecting a topic to generate a question for.

Added a 'disabled' prop to the StyledButton.tsx. This is so that you can't use the generate question button without selecting a topic first.

Added a front end error if no topic is selecting when attempting to generate a question.

* Added some simple styling to the dropdown list for selecting a topic.

* Removed any from the onClick in StyledButton.

Added an extra function for the AI call that gets passed to the onClick, instead of having to pass an arrow function directly.

* Group create & generate quesition button with styling

---------

Co-authored-by: Mark de Kraker <m.dekraker@floow.nl>
Co-authored-by: Faroeq <alves_1044@hotmail.com>